total rework of the field dict to now key on the field name instead of the field title.
Then several knock-on changes from that in how the sheet is populated.

Also removed the column assignment from the field def dict and instead
leveraged the key order of the orignal field dict to date this order.

Also reworked the model to only feature a column if the source data dict
actually has values present for that column.