Slight rework of the --incl_today option to pull data for the incomplete today and all of yesterday. In both cases each daya is pulled even if the data file already exists for these days.
In a sense, these two days are retrieves as if the --force_download was
used. But the forced download only applies to these two days.

So an invoke like:

python3 shell_em_data_util.py --days 30 --incl_today

.. will download data for the last 30 days. Data for any day that has not been
previously downloaded will be downloaded. If the same command is re-run, data
for today and yesterday will be downloaded again even though files for
both days already exist.

This suits a use case where the script is run several times a day as a
form of incremental snapshot. By pulling down todays and yesterdays data each time, we are guaranteed a complete coverage of both days