tidied up comments on Solar self-consumption and modified 'savings' to be based only on solar_credit + export_credit (both have to apply)
There is no point providing a savings is only some of the values are on
hand. This ensures that for the ESB analysis, it will show import_cost,
export_credit and bill_amount. For the likes of a Shelly/inverter setup,
it will additionally show a solar_credit and savings because of the
bigger picture of self-consumption