reworked aggregation dict generation to allow string fields to be kept as simple overwrites on each aggregation (no summation)
A time hierarchy was also added to provide a way to retain certain
time fields based on a parental hierachy. So it aggregating on day, we
would preserve the week, month and year fields. But for a monthly
aggregation, we would only keep year. This hierarchy is done using a
list and we locate the aggregation field in that list, retaining all
fields after that position. If the aggregation field is not in the time
hierarchiy list, then all time fields get removed. This caters for
aggregation on tariff, 24h or weekday usage where there is no true fixed
timepoint in play (it spans all days, months etc)