added support for default auto-filters based on the first field value.
day will force a filter on all days of the latest month
week and month will apply a filter for the latest year
and hour will apply a filter for the latest day

This will help ensure the data sets are displayed in restricted form by
default, making it less cumbersome to navigate. The full data is still
there however.