updated missing data detection to drive the population of the partial attribute in the destination file.
We can no longer work on the premise that given we're past a day, that
we can assume that days data is complete. There could be a lag for some
time after the day completes.

So it's better to ignore that we're writing data for today and just note that we encountered the missing data scenario and then mark all records in the given file as partial.