rework of refresh timer to source the value from the /data API call (refresh_interval field.
This will allow a more lazyt updating for the likes of inverters only
able to furnish updates every several minutes versus seconds with a
Shelly

Some small fixes in the chart drawing with hours and days to present
them as strings but sourced as integers from the data. The Google chart
will generalise (average?) these columns if we push them in as numbers.
For example, if we chart hour 5 and hour 6 (integers), Google can
display as 5.5 and the like. Forcing these as strings stops that
behaviour