integrated Solis module
Added support for reporting the refresh_interval (seconds) in the data
provided to the dashboard. This then drives a dynamic refresh interval
based on the specifics of the data source.

For longer refreshes, the dash will refresh every 30 seconds and
shorten this to as little as 5 seconds but no lower