rework of ESB parse into two stages: * 30-min parse (esb_dict) * hour parse (hour_dict)
This allows for parsing of multiple files with overlapping date ranges.
As a result the --file option supports a list of files to parse.

The one flaw is that the winter DST rollback will lose the composite
activity of the 2 1AM-2AM hours. Instead only the heavier of each
30-minute period across each hour will be used. Needs some thought for a
future date :)