Added support to track the acumulated power values from each string (PV1-4).
This activity is accumulated in the hourly usage record and is the sum
of all 5-min snapshot power values for each string.

Then as part of the data sweep that then converts the overall power
values to relative, this same accumulated snapshot power is used to
derive a per-string ratio. We then record the per-hour kWh, per string as a
fraction of the overall power using the ratio we derived.

It's a guess but the best one we can make and its somewhat statistically
reliable based on the N snapshots taken during the hour.

In the end, the usage records end up enhanced with four new fields..
solar_pv1, solar_pv12, solar_pv3 and solar_pv4.