fixed bug with DST issue triggering partial record scenario over and over.
The length of a day in Shelly cloud records will be 24 hours usually and
only 23 on the day that DST activates. So we use a function to determine this
special DST day and enhance the partial hour check to comparer tracked
hours against expected.

This get_hours_in_day() was previously used for the ESB parsing but now
also works well here with the Shelly API