added --solis_strings optional field to define per-string tracking for 1-8 strings. By default this is disabled.
When enabled, the resulting JSONL data will have a "solar_pv1",
"solar_pv2" etc added per separate string tracked. The existing "solar"
field will still be present and essentially the total of all the
separate PV string values reported.

the calculation is based on summing up the kW snapshot values from the
Solis 5-minute records during the hour. The hourly sum of each string is
then used to calculate its overall ratio for the kWh total generated in
that hour. So this is not bulletproof but statistically should be close
enough to the likely output per string