Dirty little trick to take advantage of the total consumed and returned values from the Shelly live API.
We now store the total_grid_import, total_grid_export & total_solar each
time the live API call is made. These are stored in the live metrics
record only.

When the cloud usage API functions are called for the last 36 hours, 30
days and 12 months, it takes a snapshot copy of that live record and
only updates it the next time the historic usage data is refreshed.

When the live data is now updated, if this snapshot live record is
available, it will calculate the delta import, export and solar since
that last cloud sync and add the differences to the most recent hour,
day and month.

The next effect is that metric records for today, this_month and
last_12_months will see trickle changes as frequently as every 10s as
the live data is being refreshed