rework of margin support to check it as a query arg inside the set_layout() function as opposed to a globl set on load.
Also added get_redirect_url(arg, value) to return a redirect URL with
the desired arg & value set. This also preserves the margin arg and any
other args that might be used going forward.

This was changed to ensure the kiosk mode with the forced ?margin=3 will
be preserved as we use the UI to cycle the layout