fixed little niggle where the live update of config would through the reload stage detecting a more fundamental change.
Soi the new logic takes a copy of the live config and applies updates from
the admin page (layout, inverter and other changes). This is then saved
to disk.

Then we live up the in-memory config from a select list of props...
dashboard and layouts only. This gives instant reaction to layout
changes, changes in metrics, timing etc.

However when the saved config is reloaded, more significant changes like
inverter, grid source etc will now be correctly detected and force a
restart if needed.