Very large rework to return to a more simplified model of building the metrics as a shared stack of cards and then have the layout and metric variations modify the shown and hidden parts of the cards, includign hiding and showing some cards based on the specific metric, like pv_total and the total.
Actual metrics like today, yesterday etc are now calculated on the fly
from the day, month and year arrays only and no longer served via the
/data API.

Still might be eventually better served with metric screens being
templates that layout cards or donuts or charts and some magic is used
to then construct these. But that is a cmplex goal to reach and manage
the scaling etc.